Extract Rpc to separate file
